# apps/daemon/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MovePageRequest": {
        "properties": {
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "position": {
            "default": "inside",
            "enum": [
              "before",
              "after",
              "inside"
            ],
            "title": "Position",
            "type": "string"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          }
        },
        "required": [
          "page_id"
        ],
        "title": "MovePageRequest",
        "type": "object"
      },
...
      "PageDocModel": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "frontmatter": {
            "additionalProperties": true,
            "title": "Frontmatter",
            "type": "object"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "path",
          "id",
          "title",
          "icon",
          "frontmatter",
          "body",
          "hash"
        ],
        "title": "PageDocModel",
        "type": "object"
      },
...
      "PageProperty": {
        "properties": {
          "colors": {
            "additionalProperties": {
              "enum": [
                "default",
                "gray",
                "brown",
                "orange",
                "yellow",
                "green",
                "blue",
                "purple",
                "pink",
                "red"
              ],
              "type": "string"
            },
            "maxProperties": 100,
            "title": "Colors",
            "type": "object"
          },
          "id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Options",
            "type": "array"
          },
          "type": {
            "enum": [
              "text",
              "number",
              "single_select",
              "multi_select",
              "date",
              "checkbox",
              "email",
              "url",
              "media",
              "status",
              "created",
              "updated"
            ],
            "title": "Type",
            "type": "string"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "id",
          "name",
          "type"
        ],
        "title": "PageProperty",
        "type": "object"
      },
...
      "PropertiesRequest": {
        "properties": {
          "base_hash": {
            "title": "Base Hash",
            "type": "string"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "properties": {
            "items": {
              "$ref": "#/components/schemas/PageProperty"
            },
            "maxItems": 100,
            "title": "Properties",
            "type": "array"
          }
        },
        "required": [
          "page_id",
          "base_hash",
          "properties"
        ],
        "title": "PropertiesRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "graite-daemon",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspace/children": {
      "get": {
        "operationId": "children_api_v1_workspace_children_get",
        "parameters": [
          {
            "in": "query",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PageDocModel"
                  },
                  "title": "Response Children Api V1 Workspace Children Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Children",
        "tags": [
          "workspace"
        ]
      }
    },
    "/api/v1/workspace/move": {
      "post": {
        "operationId": "move_api_v1_workspace_move_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovePageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDocModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move",
        "tags": [
          "workspace"
        ]
      }
    },
    "/api/v1/workspace/properties": {
      "put": {
        "operationId": "properties_api_v1_workspace_properties_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertiesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDocModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Properties",
        "tags": [
          "workspace"
        ]
      }
    }
  }
}